Clang-fix seems to be borked, reverting previous commits and fixing warnings by hand

+++ src/execution/operator/join/physical_blockwise_nl_join.cpp
@@ -197,7 +197,7 @@ public:
 public:
 	idx_t MaxThreads() override {
 		auto &sink = (BlockwiseNLJoinGlobalState &)*op.sink_state;
-		return sink.right_chunks.Count() / (static_cast<unsigned long long>(STANDARD_VECTOR_SIZE * 10));
+		return sink.right_chunks.Count() / (STANDARD_VECTOR_SIZE * 10);
 	}
 };
